+++ platform/ios/src/MGLMapView.h
@@ -105,6 +105,15 @@ typedef NS_ENUM(NSUInteger, MGLUserTrackingMode) {
     MGLUserTrackingModeFollowWithCourse,
 };
 
+typedef NS_ENUM(NSUInteger, MGLPanScrollingMode) {
+    /** The map allows the user to only scroll horizontally. */
+    MGLPanScrollingModeHorizontal               = 0,
+    /** The map allows the user to only scroll vertically. */
+    MGLPanScrollingModeVertical,
+    /** The map allows the user to scroll both horizontally and vertically. */
+    MGLPanScrollingModeDefault
+};
+
 /** Options for `MGLMapView.preferredFramesPerSecond`. */
 typedef NSInteger MGLMapViewPreferredFramesPerSecond NS_TYPED_EXTENSIBLE_ENUM;
 
@@ -688,6 +697,20 @@ MGL_EXPORT
  programmatically.
  */
 @property(nonatomic, getter=isScrollEnabled) BOOL scrollEnabled;
+
+/**
+ The scrolling mode the user is allowed to use to interact with the map.
+
+`MGLPanScrollingModeHorizontal` only allows the user to scroll horizontally on the map,
+ restricting a user's ability to scroll vertically.
+`MGLPanScrollingModeVertical` only allows the user to scroll vertically on the map,
+ restricting a user's ability to scroll horizontally.
+ `MGLPanScrollingModeDefault` allows the user to scroll both horizontally and vertically
+ on the map.
+
+ By default, this property is set to `MGLPanScrollingModeDefault`.
+ */
+@property (nonatomic, assign) MGLPanScrollingMode panScrollingMode;
 
 /**
  A Boolean value that determines whether the user may rotate the map,